Fixes tagmanager spec.
Since tagmanager is now bundled in the analytics sdk, and the download links
redirect to the new bundled sdk, the old tag manager pod nolonger works.
This spec handles the new way tagmanager is bundled.

Remark: This spec should be removed in favor of a subspec on the google analytics spec.
The reason is that the libGoogleAnalyticsServices.a will be installed twice if an app
requires both analytics and tagmanager. However that is very unlikely. Instead of having
a broken tagmanager spec, this is better then nothing.